feat: add useDebounce hook for improved input handling
