+++ packages/api-v2/src/hooks/index.ts
@@ -1,4 +1,5 @@
 export { default as useAuthorize } from './useAuthorize';
 export { default as useBalance } from './useBalance';
+export { default as useDebounce } from './useDebounce';
 export { default as useRemoteConfig } from './useRemoteConfig';
 export { default as useServerTime } from './useServerTime';
