+++ packages/trader/src/AppV2/Components/TradeParameters/Barrier/barrier-input.tsx
@@ -1,27 +1,11 @@
 import React from 'react';
 import { observer } from 'mobx-react-lite';
 
+import { useDebounce } from '@deriv/api-v2';
 import { ActionSheet, Chip, Text, TextField, TextFieldAddon } from '@deriv-com/quill-ui';
 import { Localize, useTranslations } from '@deriv-com/translations';
 
 import { useTraderStore } from 'Stores/useTraderStores';
-
-// Debounce hook for real-time validation
-const useDebounce = (value: string, delay: number) => {
-    const [debouncedValue, setDebouncedValue] = React.useState(value);
-
-    React.useEffect(() => {
-        const handler = setTimeout(() => {
-            setDebouncedValue(value);
-        }, delay);
-
-        return () => {
-            clearTimeout(handler);
-        };
-    }, [value, delay]);
-
-    return debouncedValue;
-};
 
 const chips_options = [
     {
